+++ src/index.js
@@ -1,0 +1,6 @@
+// Components
+export { default as ProfileRules } from './ProfileRules'
+export { default as ProfileContainer } from './ProfileContainer'
+
+// Helper Functions
+export { addValidation, removeValidation } from './validateProfile'