+++ src/inputs/StyleguideInput/index.js
@@ -2,10 +2,10 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { intlShape, injectIntl } from 'react-intl'
 import Input from '@vtex/styleguide/lib/Input'
-import RuleFieldShape from './propTypes/RuleFieldShape'
-import ProfileFieldShape from './propTypes/ProfileFieldShape'
+import RuleFieldShape from '../../propTypes/RuleFieldShape'
+import ProfileFieldShape from '../../propTypes/ProfileFieldShape'
 
-const ProfileInput = ({ field, intl, data, onChange }) => {
+const StyleguideInput = ({ field, intl, data, onChange }) => {
   return (
     <div
       className={`vtex-profile-form__${field.name} ${
@@ -35,7 +35,7 @@ const ProfileInput = ({ field, intl, data, onChange }) => {
   )
 }
 
-ProfileInput.propTypes = {
+StyleguideInput.propTypes = {
   /** Rules for the field this input represents */
   field: RuleFieldShape.isRequired,
   /** Data this input will display */
@@ -46,4 +46,4 @@ ProfileInput.propTypes = {
   intl: intlShape.isRequired,
 }
 
-export default injectIntl(ProfileInput)
+export default injectIntl(StyleguideInput)